+++ ronr_web_app/src/App.js
@@ -3,8 +3,8 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import LoginPage from './components/login-pages/login-page'; // Start page
 import Registration from './components/login-pages/registration'; // Registration page
 import PasswordReset from './components/login-pages/password-reset'; // Password reset page
-import HomePage from './components/home-page'; // Home page
-import ChatPage from './components/chat-page';
+import HomePage from './components/home-page/home-page'; // Home page
+import ChatPage from './components/discussion-page/chat-page';
 
 function App() {
   return (
